[REQUIRES REVISION] Fix issue in UpdateEmotion
ISSUE: Occasionally missed an appraisal because _lastFrameAppraisal was
the same of frame.LastChange. Test scenario was the three consecutive events for the
trick end in sueca (Dialogue(State),Trick(Winner),Trick(End)) and sometimes it
missed the appraisal of the last event, even when found a match in the set of appraisal
rules.

I think it happened only once in a while depending on the number of updates in the
rpc during the appraisal of the events. Also because when increasing the
Console.WriteLines in the middle of appraisal rules evaluation (that
forced the execution to last longer), the number of missed
appraisals decreased and was harder to reproduce the bug.

Me and @jovemengenheiro looked into this issue and solved it in this
commit. At least for the sueca scenario is working properly. However it
needs revision, specially for other scenarios that we might not be
considering.